#184 Fix astropy module shadowing by renaming astropy.py to astropy_utils.py
