+++ neclib/core/normalization/__init__.py
@@ -1,5 +1,5 @@
 """Collection of type converters which allow type-flexible read-in."""
 
 from .array import NPArrayValidator  # noqa: F401
-from .astropy import QuantityValidator, get_quantity  # noqa: F401
+from .astropy_utils import QuantityValidator, get_quantity  # noqa: F401
 from .function_defaults import partial  # noqa: F401
